+++ wattpad_scraper/utils/request.py
@@ -19,6 +19,7 @@
 import requests
 from fake_headers import Headers
 import atexit
+import os
 
 headers = Headers(
     browser='chrome',
@@ -26,12 +27,28 @@ headers = Headers(
     headers=True
 )
 
+def login(username,password):
+    login_endpoint  ="https://www.wattpad.com/login?nextUrl=%2Fhome"
+    data = {
+        "username": username,
+        "password": password
+    }
+    session.post(login_endpoint,data=data,headers=headers)
 
 headers = headers.generate()
 response_memory = {}
 
 session = requests.Session()
+USER_LOGGED_IN = [False]
+
+
+
 def get(url):
+    if not USER_LOGGED_IN[0]:
+        if "WATTPAD_USERNAME" in os.environ and "WATTPAD_PASSWORD" in os.environ:
+            login(os.environ['WATTPAD_USERNAME'],os.environ['WATTPAD_PASSWORD'])
+        USER_LOGGED_IN[0] = True
+
     if url not in response_memory:
         response_memory[url] = session.get(url,headers=headers)
     return response_memory[url]