+++ CHANGELOG.md
@@ -75,3 +75,16 @@ Per thread removed.
 - Search books
 - Get book by url
 - Convert books to epub
+
+## [0.0.24]
+Auth System (Beta)
+### Major Changes:
+- Added Auth System
+### Minor Changes:
+- Parse book chapters content again if chapter content broken or empty.
+
+### Features
+- Search books
+- Get book by url
+- Convert books to epub
+- Login to wattpad